requirements.txt
aniso8601==9.0.1
attrs==21.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.0.3
converge==1.0.2
coverage==6.2
cracklib==2.9.3
dictpath==0.1.3
flake8==4.0.1
Flask==2.0.2
Flask-Cors==3.0.10
flask-restx==0.5.1
gunicorn==20.1.0
idna==3.4
iniconfig==1.1.1
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==3.2.0
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
more-itertools==8.14.0
openapi-core==0.14.2
openapi-schema-validator==0.2.3
openapi-spec-validator==0.4.0
packaging==21.3
parse==1.19.0
passlib==1.7.4
pathable==0.4.3
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
pytest-cov==3.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-mock==3.6.1
python-ldap==3.3.1
pytz==2021.3
PyYAML==6.0
requests==2.28.1
six==1.16.0
toml==0.10.2
tomli==2.0.0
typing_extensions==4.3.0
urllib3==1.26.12
waitress==2.0.0
Werkzeug==2.0.2
